custom-config/requirements.txt
aiohttp>=3.8.0
PyYAML>=6.0
types-PyYAML>=6.0
typing-extensions>=4.5.0 